oracle function and get_max_amplitude function
Oracle function.
Inputs: amplitudes list. a list of amplitudes of the corresponding states.
Returns: list of amplitudes with the one we want flipped by sign.

get_max_amplitude.
Inputs: amplitudes list.
Returns: the maximum amplitude and it's index in the amplitudes list (so corresponding state).

IMPORTANT: My labels start from 1 not 0. So might need to change this but it's an easy fix.