Incorporated uncertainty
Added method observe() into state class. This method uses the amplitudes of the states to "collapse" the superposition of the states into one, depending on how probable each state is.

This in principle is done using a simple random float generation between 0 and 1. The actual method is a bit more involved and may not be clear at first glance. More info up close