Changed names of files. Tidied up
All runs can be done at run.py.

grovers.py contains all the functions needed to run the algorithm.

The rest is all the same, just changed the names of classes/methods and comments, so that now:
state = definite ... state,
register = the system (i.e. superposition of states)

The previous naming system was creating confusion and was most probably wrong